pdf-chat/requirements.txt
docx2txt==0.8
PyPDF2
python-dotenv
faiss-cpu==1.7.4
langchain==0.0.340
openai==1.3.5
openapi-schema-pydantic==1.2.4
pandas==2.1.3
requests==2.31.0
streamlit==1.28.2
tenacity==8.2.3
tiktoken==0.5.1
azure-identity==1.15.0
azure-search-documents==11.4.0b8
streamlit-agraph==0.0.45
gremlinpython==3.7.1
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
